feat: Add UpdateRole endpoint to RoleController
This commit adds the UpdateRole endpoint to the RoleController in the Dev.Plugin.Authen.Core.Controllers namespace. The UpdateRole endpoint allows updating a role by sending an HTTP PUT request with the updated role information in the request body. The UpdateRoleCommand is used to encapsulate the updated role data. The role is updated by calling the _mediator.Send() method with the UpdateRoleCommand as the parameter. After the role is updated, the endpoint returns a NoContent response.

Note: This commit also includes changes to the UpdateRoleCommandValidator in the Dev.Plugin.Authen.Core.UseCases.Roles namespace. The RoleNameUniqueAsync method is updated to check if the role name is unique by searching for roles with the same name and different IDs.